+++ public/src/js/controllers/scanner.js
@@ -1,7 +1,7 @@
 'use strict';
 
 angular.module('insight.system').controller('ScannerController',
-  function($scope, $rootScope, $modalInstance, Global) {
+  function($scope, $rootScope, $uibModalInstance, Global) {
     $scope.global = Global;
 
     // Detect mobile devices
@@ -96,7 +96,7 @@ angular.module('insight.system').controller('ScannerController',
 
     var _scanStop = function() {
       $scope.scannerLoading = false;
-      $modalInstance.close();
+      $uibModalInstance.close();
       if (!$scope.isMobile) {
         if (localMediaStream.stop) localMediaStream.stop();
         localMediaStream = null;
@@ -112,7 +112,7 @@ angular.module('insight.system').controller('ScannerController',
     qrcode.callback = function(data) {
       _scanStop();
 
-      var str = (data.indexOf('bitcoin:') === 0) ? data.substring(8) : data; 
+      var str = (data.indexOf('dash:') === 0) ? data.substring(8) : data;
       console.log('QR code detected: ' + str);
       $searchInput
         .val(str)
@@ -124,7 +124,7 @@ angular.module('insight.system').controller('ScannerController',
       _scanStop();
     };
 
-    $modalInstance.opened.then(function() {
+    $uibModalInstance.opened.then(function() {
       $rootScope.isCollapsed = true;
       
       // Start the scanner